Revert "created a file structure"
